Add bathrooms float support with NUMERIC(3,1) and Pydantic validation
- models/listing.py: bathrooms column changed from Integer to Numeric(3,1)
- schemas/listing.py: Bathrooms type alias (Annotated[float, Field(ge=0,
  multiple_of=0.5)]); ListingCreate + ListingUpdate use Optional[Bathrooms];
  ListingResponse uses Optional[float]
- alembic migration d5e6f7a8b9c0: INTEGER -> NUMERIC(3,1) via postgresql_using
  cast; chains from b1c2d3e4f5a6; includes downgrade path

+++ models/listing.py
@@ -19,7 +19,7 @@ class Listing(Base):
     price = Column(Numeric(12, 2), nullable=False)
     location = Column(Text, nullable=False)
     bedrooms = Column(Integer)
-    bathrooms = Column(Integer)
+    bathrooms = Column(Numeric(3, 1))
     area_sqft = Column(Integer)
     lot_size_sqft = Column(Integer)
     roi = Column(Numeric(5, 2))
